FVCalc

+++ FVCalc.java
@@ -1,9 +1,12 @@
 // Computes the future value of a saving investment.
+
+import java.util.stream.Gatherer.Integrator;
+
 public class FVCalc {
 	public static void main(String[] args){
-		double saving = Double.parseDouble(args[0]);
-		double interestRate = Double.parseDouble(args[1]);
-		double timeFrame = Double.parseDouble(args[2]);
+		int saving = Integer.parseInt(args[0]);
+		int interestRate = Integer.parseInt(args[1]);
+		int timeFrame = Integer.parseInt(args[2]);
 		double rate = interestRate / 100;
 		double futureValue = saving;
 		double time = timeFrame;
